1.修改了考试界面；添加了提交按钮 2.修改了上一题下一题按钮的显示方式 3.修改了答案提交的方式

+++ 源码/TestSystem/src/main/java/com/examination/cotroller/LoginCotroller.java
@@ -143,7 +143,7 @@ public class LoginCotroller {
 		String userPhone = request.getParameter("phone");
 		String usernumcard = request.getParameter("numcard");
 		String useremail = request.getParameter("email");
-		Integer user_no = Integer.valueOf(userno);
+		Long user_no = Long.valueOf(userno);
 		User user = new User(userId, userCode, pwd1, userName,user_no,usernumcard, userPhone,useremail);
 		isadd=ud.addUser(user);
 		request.setAttribute("REGISTER", isadd);
